refactor(F-626): perform small fixes
