+++ frontend/src/app/[locale]/(main)/dashboard/components/DashboardTable.tsx
