--- frontend/src/app/[locale]/(main)/history/[id]/components/HistoryLabels.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import { useTranslations } from 'next-intl';
-
-export function StatisticsLabel() {
-  const t = useTranslations('History');
-  return <div className="text-xl font-bold text-bw-70">{t('scores')}</div>;
-}
-
-export function PreviousSessionsLabel() {
-  const t = useTranslations('History');
-  return <div className="text-xl font-bold text-bw-70">{t('previousSessions')}</div>;
-}
